Refactor: extracted a few parsers out of OasTypeParser
